refactor: Mapper 위치 수정
