+++ android/app/src/main/java/com/woowacourse/friendogly/remote/source/FootprintDataSourceImpl.kt
@@ -1,9 +1,9 @@
 package com.woowacourse.friendogly.remote.source
 
+import com.woowacourse.friendogly.data.mapper.toData
 import com.woowacourse.friendogly.data.model.FootprintInfoDto
 import com.woowacourse.friendogly.data.source.FootprintDataSource
 import com.woowacourse.friendogly.remote.api.FootprintService
-import com.woowacourse.friendogly.remote.mapper.toData
 
 class FootprintDataSourceImpl(private val service: FootprintService) :
     FootprintDataSource {
